day30 add sth about txn

+++ src/default.rs
@@ -4,7 +4,7 @@ use std::mem::size_of;
 
 // use crate::skl::Node;
 
-pub(crate) const MAX_VALUE_THRESHOLD: i64 = 1 << 20;
+pub(crate) const MAX_VALUE_THRESHOLD: usize = 1 << 20;
 pub(crate) const DEFAULT_DIR:&str="./tmp/badger";
 pub(crate) const DEFAULT_VALUE_DIR:&str="./tmp/badger";
 pub(crate) const LOCK_FILE:&str="LOCK";
